修改数据库id类型为uint32，添加DeleteModel删除标记

+++ service/app/account/internal/biz/biz.go
@@ -12,13 +12,13 @@ var ProviderSet = wire.NewSet(NewAccountUseCase)
 type AccountRepo interface {
 	// CreateEMailAccount 创建邮箱账户
 	// 返回创造账户的id
-	CreateEMailAccount(context.Context, *Account) (uint64, error)
+	CreateEMailAccount(context.Context, *Account) (uint32, error)
 
 	// GetAccountByUserID 通过用户ID获取账户
-	GetAccountByUserID(ctx context.Context, userid uint64) (*Account, error)
+	GetAccountByUserID(ctx context.Context, userid uint32) (*Account, error)
 
 	// GetAccountByID 通过Account主键ID获取账户
-	GetAccountByID(context.Context, uint64) (*Account, error)
+	GetAccountByID(context.Context, uint32) (*Account, error)
 	// GetAccountByEMail 通过用户邮箱获取账户
 	GetAccountByEMail(context.Context, string) (*Account, error)
 	// GetAccountByPhone 通过用户手机号获取账户
@@ -41,7 +41,7 @@ type AccountRepo interface {
 	UpdateAccount(context.Context, *Account) error
 
 	// BindUser 绑定用户
-	BindUser(context.Context, uint64, uint64) error
+	BindUser(context.Context, uint32, uint32) error
 }
 
 type AccountUseCase struct {
